+++ app/dashboard/_components/TextareaSection/index.tsx
@@ -1,0 +1,7 @@
+export default function TextareaSection() {
+  return (
+    <div className="h-[520px] overflow-y-hidden rounded bg-white px-4 py-2">
+      <textarea className="h-full w-full border border-gray-200" />
+    </div>
+  );
+}
