+++ components/Tab/context/index.ts
@@ -1,0 +1,8 @@
+import { Dispatch, SetStateAction, createContext } from "react";
+
+type TabContext = {
+  selectTab: string;
+  setSelectTab: Dispatch<SetStateAction<string>>;
+};
+
+export const TabContext = createContext<TabContext | null>(null);
