+++ components/Tab/index.tsx
@@ -1,0 +1,3 @@
+export { default as TabContainer } from "./TabContainer";
+export { default as TabList } from "./TabList";
+export { default as TabContents } from "./TabContents";
